requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.24.0
scikit-image>=0.21.0
matplotlib>=3.7.0
joblib>=1.3.0
streamlit>=1.28.0
opencv-python>=4.8.0
Pillow>=9.5.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
pdf2image>=1.16.3